Extract clutter to utility class
